writing to DB now --GoBack to this if DB does not work
